[web] Move App's content to an Overview component
